Move vercel tests to workspace
