+++ packages/fika-collect-vercel/test/fixtures/schema.json
