+++ packages/fika-collect-vercel/test/util/submitSurveyRequest.test.ts
@@ -2,7 +2,7 @@ import { readFileSync } from 'node:fs';
 import { dirname, join } from 'node:path';
 import { fileURLToPath } from 'node:url';
 import { describe, it, expect } from 'vitest';
-import { surveySubmissionRequestSchema } from '../../../api/util/requestSchema.js';
+import { surveySubmissionRequestSchema } from '../../../../api/util/requestSchema.js';
 
 const __dirname = dirname(fileURLToPath(import.meta.url));
 const responseFixture = JSON.parse(
